added code for switching colors in canvas.tsx
