+++ client/src/utils/canvas.tsx
@@ -1,7 +1,112 @@
+import { useRef, useState, useEffect} from "react"
 
-//should update the page depending on which pixel was clicked on.
-function canvas (){
-    
+const colors: {} = {
+    red: '#ff0000',
+    black: '#000000',
+    greyblue: '#2980B9',
 }
 
-export default canvas
+let selectedColor: string  = '#000000';
+
+//should update the page depending on which pixel was clicked on.
+const CanvasComponent: React.FC = () => {
+    const canvasRef = useRef<HTMLCanvasElement | null>(null);
+    const canvasWidth = 500;
+    const canvasHeight = 500;
+    const pixelSize = 10;
+
+    const [isPainting, setPainting] = useState(false);
+
+    const changeColor = (color: string) => {
+        return selectedColor = color;
+    }
+
+    //draws an initial grid onto the canvas.
+    const drawGrid = (ctx: CanvasRenderingContext2D) => {
+        for (let x= 0; x< canvasWidth; x += pixelSize){
+            for(let y=0; y < canvasHeight; y += pixelSize){
+                ctx.fillStyle = '#ffffff';
+                ctx.fillRect(x, y, pixelSize, pixelSize);
+
+                ctx.strokeStyle = '#cccccc';
+                ctx.strokeRect(x, y, pixelSize, pixelSize);
+            }
+        }
+    };
+
+    const drawPixel = (ctx: CanvasRenderingContext2D, x: number, y: number) => {
+        //grabs the mouse click coordinates
+        const pixelX = Math.floor(x /pixelSize) * pixelSize;
+        const pixelY = Math.floor(y /pixelSize) * pixelSize;
+
+        const currentColor = ctx.getImageData(pixelX, pixelY, 1, 1).data;
+        const isRed = currentColor[0] === 255 && currentColor[1] === 0 && currentColor[2] === 0;
+
+        //between red and black currently
+        ctx.fillStyle = selectedColor;
+        ctx.fillRect(pixelX, pixelY, pixelSize, pixelSize);
+    }
+
+    const handleMouseDown = (event: React.MouseEvent<HTMLCanvasElement>) => {
+        const canvas = canvasRef.current;
+        if(!canvas) return;
+
+        const ctx = canvas.getContext('2d');
+        if(!ctx) return;
+
+        // Grabs the canvas's bounding retangle and the mouse click Coordinates
+        const rect = canvas.getBoundingClientRect();
+        const x = event.clientX - rect.left;
+        const y = event.clientY - rect.top;
+
+        drawPixel(ctx, x, y);
+        setPainting(true);
+    }
+
+    const handleMouseMove = (event: React.MouseEvent<HTMLCanvasElement>) => {
+        if(!isPainting) return;
+
+        const canvas = canvasRef.current;
+        if(!canvas) return;
+
+        const ctx = canvas.getContext('2d');
+        if(!ctx) return;
+
+        const rect = canvas.getBoundingClientRect();
+        const x = event.clientX - rect.left;
+        const y = event.clientY - rect.top;
+
+        drawPixel(ctx, x, y)
+    }
+
+    const handleMouseUpLeave = () => {
+        setPainting(false);
+    };
+
+    useEffect(() => {
+        const canvas = canvasRef.current;
+        if(!canvas) return;
+
+        const ctx = canvas.getContext('2d');
+        if(!ctx) return;
+
+        drawGrid(ctx);
+    }, []);
+
+    return (
+        <div>
+            <div className="canvasHolder">
+                <canvas id='paintMain' ref={canvasRef} width={canvasWidth} height={canvasHeight} onMouseDown={handleMouseDown} onMouseMove={handleMouseMove} onMouseUp={handleMouseUpLeave} onMouseLeave={handleMouseUpLeave} style={{border: '1px solid red'}}></canvas>
+            
+            </div>
+            <div className="colorSelectors">
+                <button style={{backgroundColor: '#ff0000'}} onClick={() => changeColor('#ff0000')}></button>
+                <button style={{backgroundColor: '#2980B9'}} onClick={() => changeColor('#2980B9')}></button>
+                <button style={{backgroundColor: '#000000'}} onClick={() => changeColor('#000000')}></button>
+            </div>
+        </div>
+        
+    )
+}
+
+export default CanvasComponent
